license-server/requirements.txt
fastapi
uvicorn
sqlalchemy
cryptography
pydantic
passlib[bcrypt]
jinja2
python-multipart
python-jose[cryptography]
bcrypt==4.0.1